Fix surround sound support
A change in ffmpeg v4 led to a change in the ADTS output structure for
5.1 surround sound.  It seems that the channel layout changed from 5.1
to 5.1(side).  This change corresponds to its triggering of a Packager
issue in which the ADTS structure is not supported.  To work around
this and packager 6-channel audio, explicitly set the channel format
in ffmpeg.

See also google/shaka-packager#598, where this was discussed.

The solution was found on https://trac.ffmpeg.org/ticket/6974 .

Closes #6

Change-Id: Ief88bdcae9756d0c264ec36ee1160940193cfead