Delete the discard node
This is not being used, and it seems not to work well.  Tail -f seems
to read from each pipe in turn, but not "fairly".  Instead, when a
pipe has bytes, it reads from that pipe until it blocks.  If that pipe
never blocks, the others will never be read from.  This has the effect
that if the audio, for example, never stops, then it will never read
any video.  And of course, if you're debugging a live stream, the
audio will never stop.

Change-Id: I99526a139bab99edb1faede400a16ac4f337249c