Put inputs first, use map unconditionally
The arguments for all inputs now come first, before any of the outputs.
This is mostly helpful in understanding the generated command line,
but doesn't fix any known bug.

This also uses -map unconditionally to map inputs to outputs instead of
only for VOD, and fixes the -map argument to use the input file number
instead of 0.  This could fix bugs, although we don't have any known
issues filed about this.  A regression test has been added for this.

Finally, to clarify the code a bit, this renames processed_inputs in
controller_node and _inputs in transcoder_node to input_paths.  Since
transcode_node relies on the config's inputs being equal in length to
the input paths, it also asserts that this is true.

Change-Id: I0083ac2ceb605a4db3e762bf87decccc48eeb2e0