Tweak defaults, examples, and explanations
This removes some unnecessary features from sample configs, improves
the explanations of certain fields, and changes the default segment
size to 4 seconds.  The previous default of 10 would have been too
large for a live stream, and it seems best to have a default that
size that would work for both VOD and live.

Change-Id: I051585577b3f0080821dfd7858f323b862327cf4