Refactor handling of stop and is_running.
This makes it so we stop all the subprocesses on exit.  First, this
uses a context manager on the controller to stop it when we leave the
block.  Then this also adds a __del__ method to cleanup when we shutdown
if we can.

This also changes how we track the process.  This adds an Errored state
in addition to whether it is running.  This allows us to catch when they
exit with a non-zero exit code.  Now if one of the subprocesses exits
with an error, we will return an error.

Fixes #20

Change-Id: I74bbf54f802648c81d8eea5df999d4648bc424b0