Handle looped inputs directly in TranscoderNode
It turns out that the options to loop an input file are part of
FFmpeg's input options (meaning they apply to each input
individually).  This means we can handle looped inputs directly in
TranscoderNode and drop LoopInputNode completely.  This simplifies the
pipeline to a single instance of FFmpeg in all cases.

This is not only simpler, but seems to be more efficient, since we
aren't running multiple full copies of the looped input through
multiple pipes, only to have some of the individual streams in those
pipes discarded by TranscoderNode's FFmpeg instance.

This also simplifies a few other checks and if/elif/elif/... blocks
based on the removal of LoopInputNode and cleanup around
InputType.looped_file.

Change-Id: I0b6d2ae9f4130e7ec40eecba823943717578b8a3