Remove redundant hardware encoding section
The hardware encoding section in the overview doc made more sense when
it was part of README.md.  But now that we have generated docs with a
TOC linking directly to the hardware encoding doc, this is redundant.
So just mention hardware encoding once in the features list, and let
the hardware encoding doc stand alone without a link from overview.

Change-Id: I629be9a067358bf8ff218dc4d7434716ca4405af