Autodetect frame_rate, resolution, is_interlaced, track_num
For most types of inputs, you no longer need to specify the input
frame rate or resolution, and we can now detect interlaced video
automatically.  Furthermore, track_num is now per media_type, meaning
that a default of 0 can be used to mean the first track of that type,
without the need to know an explicit, overall track number from the
input file.

Change-Id: I7ca7f9a34b4f7f9a0218241609dc311103691dd4