Clean up H264 arguments
There is no longer any reason to avoid B-frames, since Chrome has
fixed its DTS/PTS issues.  So drop '-bf 0' from the transcoding
command-line.

The H264 level argument does not need to be set manually.  FFmpeg will
look up the correct level based on the resolution, frame rate, and
bitrate, so setting this is redundant at best and harmful at worst (if
we set one that is too low, tricking a device into playing a stream it
can't handle).  Dropping the level argument allows us to only worry
about the profile, which must still be set explicitly (or else it
defaults to "main").

Change-Id: Ide0fbbf6ec8e6e6b703c0056ea7637b5f64f266c