Fix language setting in PackagerNode
The language setting in Shaka Packager rejects "und" as a language.
This used to be screened out until the custom bitrates feature was
added.  This bug does not affect any release versions.

This change fixes PackagerNode to exclude "und" from the Packager
language settings.  Shaka Player will supply "und" as a language if
there is no language metadata in the manifest/playlist.

Discovered while investigating issue #34

Change-Id: Ic3a479466a1a1944b7b73f919044aeed635ed1fe