Make common errors easier to read
The most common errors someone will see are VersionError (dependency
missing or not up-to-date) and ConfigError (mistake in a config file).

This makes those easy to read by catching them and only printing the
message without the Python backtrace.  This should make it easier for
users to fix common issues.

Change-Id: I6a599e3b256f9626e1656bc9eb9f68011e61bdc4