Fixed rounding errors in width in HLS playlist
When scaling the input, FFmpeg calculates the ideal width for each
output resolution for us.  This can lead to small rounding errors
which show up in the Sample Aspect Ratio (SAR).  An SAR which is not
1:1 seems to affect DASH & HLS outputs differently in Shaka Packager.
In DASH, width is correct, and SAR is an unusual value like 5120:5123.
In HLS, the width is off by one, causing playback failures in
ExoPlayer.

Closes #36

Change-Id: I874c3bfd83206b7c88384a2364234a284e8294a7