Reduce the minimum version of gcloud
The matches the version available in Ubuntu 16.04 in Google Compute
Engine.  The bug fixes introduced in gsutil 4.40 are nice to have, but
the fixes introduced in gsutil 4.33 are more critical.  So this older
version is good enough, and gets us compatibility with this build in
GCE.

See the release notes for gsutil for details:
https://github.com/GoogleCloudPlatform/gsutil/blob/master/CHANGES.md

Change-Id: I7c79e4063b2317022946ddf84d6a6f37befd4388