Add type info to RuntimeMap
This makes RuntimeMap a Generic base class which contains a map from
string keys to the various subclass instances associated with those
keys.

This is a big refactor of this type, which will now be well understood
by mypy.  As part of this, it no longer tries to mimic most of the
behaviors of Enum types.

This also removes the separate map types and moves all the class-level
mapping methods into the config classes which define the values.  For
example, instead of Resolution being a map from keys to
VideoResolution config objects, the map is now stored in the
VideoResolution class itself, simplifying the type info maintained.

Finally, this makes distinct "Name" types for these mapped types.  The
mapped type itself is always a value (full config object), whereas the
name is a string which validates against the valid names from the
config.

Change-Id: I7d59b44381b8c05495e43e4c06c4968d14b397f7