Fix duplicate transcoder outputs with multiple audio languages
When encoding "A" audio streams into "B" audio codecs, we should end
up with "A*B" audio outputs.  We generated the correct number of
output objects, but a bad filtering condition in the TranscoderNode's
inner loop caused us to start "A^2*B" encoders.

There were the correct number of pipes between TranscoderNode and
PackagerNode, and the correct number of output streams on disk, so
this is not detectable in an end-to-end test.  But without this fix,
this multi-audio-language scenario uses lots of extra CPU to encode.

Change-Id: I3b87e2c0cce7cd22eca2be5ea14cfde6b16a93bc