Fix test flake
We had a test that was only working with older versions of Shaka
Player.  Our test dependency was anything newer than v2.5.5, but a bug
fix in Shaka Player to correctly skip WebM in HLS broke some tests
that produced WebM in HLS.  There is no reason that the test should
have used WebM, so this changes it to AAC to make sure it works
correctly in HLS and in all versions of Shaka Player.  This also
updates the Shaka Player dependency to the latest v2.5.10 release to
ensure consistent results.

Change-Id: Ieb8ebadf60f1806586fe3d79d03ee56a94cf2880