Fix mypy 0.770 errors
After upgrading to mypy 0.770, two type errors came up that were not
detected with earlier versions.  In both cases, mypy is being strict
in a scenario where it lacks sufficient information to verify the
types.

In one case, Field.get_subtypes could not verify that all sources of
type information conformed to the return type.  Because there is some
meta-programming going on here, and there are sufficient comments to
explain where the info comes from and conditions on its use, the
solution is to cast the type in the return statement.

In the other case, we got complaints about RuntimeMap being an
abstract class.  I could not understand what the issue was with this,
but a simple solution is to remove the @abc.abstractmethod decorator
and rely on a runtime exception if the one abstract method is not
overridden by subclasses.

The error was "Only concrete class can be given where
Type[RuntimeMap[RuntimeMapSubclass]] is expected", and it pointed to
the class definition, which I found confusing.  This may have referred
to the Generic typing which references RuntimeMapSubclass, but it's
unclear how else I should resolve this.  Given that this is more
low-level metaprogramming, and the type-safety isn't compromised by
making the class concrete, this seems like the best trade-off between
correctness and effort.

Change-Id: Iefed065fc80b97b67cbbe48e959098c3ab5eebe2