test: Fix test failures with newer Flask version
A Flask upgrade caused test failures in which the output folder could
not be read by the browser.  This is because Flask overrode our route
for output_files with its own static_folder handler.  The fix is to
remove the static_folder handler, which we didn't need in the first
place.

Change-Id: I024c4194c02971dd82d5fedbb94632b6f8d0c6a6