Fix packaging failures with long-running content
Some long-running VOD content has been failing since we switched to
using the MP4 container in the pipeline, as part of AV1 support.  It
turns out that this is because we did not specify a fragment duration
for FFmpeg.  We had expected that the frag_keyframe flag would be
enough to introduce an output fragment per keyframe, but that does not
seem to be the case.  That flag only seems to affect video.  To force
fragmentation of audio as well, we need an explicit fragment duration.
The ensure that the output of FFmpeg doesn't contain single large MP4
boxes that Shaka Packager can't consume from a pipe.

It also seems that the faststart flag was voodoo that did not affect
anything, so it has been removed.

Because this issue only seems to appear with long-running or larger
content, we can't easily add a regression test without large inputs
and much longer test runs.

Closes #64

Change-Id: Ie92ec1ca07f4d75c61eb3e95109e139cf20b470f