test: Codec tests no longer depend on browser support
We no longer have to rely on browser support for a codec to check
Streamer support for a codec.  We can check the manifest right after
parsing, and ignore any playback errors.

This also adds a test case for PR #74 (HEVC support) which was too
difficult to add before this refactor.

Change-Id: I37edb8ff7c94134e74e7a86f85c5b44bf4b38e24