using enum._missing_ hook to generically handle the hw: variants
This PR removes the hw: enum variants making the enum only for the base codec variants.

The if a codec is prefixed with 'hw:', it will be caught in the _missing_ hook, we will try to know if it can be one of the base variants or not, if it is a base variant we will set the _hw_acc to True for this codec, if not we will fall back the the super()._missing_ so to have a proper error message by the `configuration.py`.