docs: Fix missing members in generated docs
The docs were meant to include details of the config fields and the
module API, but at some point, those quit working.  I suspect it was
due to an update in sphinx or its modules, but I am not sure, as I use
auto-updated packaged versions from Debian.

The current versions of autoclass/automodule require default options
to be specified differently.  This fixes those default options.

Also, in our docs post-processing callback, a change was needed in how
we get the full names of the members we are documenting.

Change-Id: I50b6af7ebd2c23f26c55abbd97e3336d8665f24f