docs: Auto-link to types in config format docs
Every type defined in these docs has a hash that can be linked to to
see its fields or enum values.  But Sphinx isn't linking to them
automatically.  This may be caused by a missing Sphinx plugin, but
since we heavily post-process the generated type info for our config
docs (see docs/source/conf.py), it is easiest to add these links in
JavaScript after the docs load.

Now where the docs say something like "inputs: list of Input", the
word "Input" is a link to the definition of that type.

Change-Id: I148fa76de01baa53d235944d66c1fbcf74cd5ce6