Use enum._missing_ hook to generically handle "hw:" codecs (#86)
This PR removes the "hw:" enum values, making the enum only for the base codecs.

If a codec is prefixed with "hw:", it will be caught in the _missing_ hook.  We will try to know if it can be one of the base variants or not.  If it is a base variant, we will set the _hw_acc to True for this codec.  If not, we will fall back to super()._missing_ to have a proper error message.