Major change for the concat_aud_common
Before the `Channel count as an input feature` PR or rather `Channel layout as an input feature` shall i say, we had this problem where we are not guaranteed that a lower channel count exists for every input.  For example, an input(period) might have six channel and no two channel, while another input might have two channels AND six channels, though in this configuration we have 2 & 6 channels but we can't generate 2 playlists (2 & 6), what we can do instead in this case is to generate two playlists, one has (2 and 6 channels) and one having (6 and 6 channels), And we have to mark both playlists with `CHANNELS="6"` because both of them has a maximum of 6 channel count across all periods, and also since the audio playlist is a media playlist, so no way for shaka player to know the bitrate for each playlists, so shaka player will just look at these playlists as identical.

With the new change, we are guaranteed to have the lower channel for some input if a higher one exist (e.g if an input has 6 channels so it must have 2 also).

This way we can set the `CHANNELS="val"` attribute appropriately and we guarantee that we will generate a playlist for each unique channel layout we have in the inputs.