ci: Workflow for building and testing PRs (WIP)
This will build and test Shaka Streamer on various OS and Python
versions when a PR is created or updated.

I still get flakey test results in GitHub's environment that I can't
replicate, and they span OS and python version.

There is more CI work to do here, but this is a good start.

Change-Id: I5e9fb3fc8786d32744cff62d107ddb8d4e250ed5