build: Fix another Python 3.6 type-info issue
This fixes another issue with runtime type reflection in Python 3.6.

Type info was missing for certain enums, but it was difficult to find.
The tests would simply fail with "500 INTERNAL ERROR" and no useful
info.

This improves the error output so that we can see it was a failed
assertion, and it improves the assertion (field.type is not None) to
indicate which field was missing type info.

Finally, this fixes the missing type info for enums in Python 3.6.

Change-Id: I55d15b298a2d919de3798100554c475d4b8e3cb8