cleanup: Refactor version checks
The binary module and main app may get bugfix updates on a completely
different cycle than one another, so we should only check that the
first two components of their version numbers match.

This also renames the version check function for external commands
(for clarity) and creates a more structured constructor for
VersionError.

Finally, this aligns the shaka-streamer version number with the first
release of the binary package, so that it passes the new check.

Change-Id: Ia57b34139e5c1b148eb95bc887924aa143bf9731