build: Fix usage of local streamer_binaries module
The old code would search the directory for the binary to use.  This
worked fine if the package were installed, but when adding the module
path to PYTHONPATH, this technique would fail because the folder would
have executables for all architetures.

Now we will compute the exact filename we expect for each exectuable,
allowing the module to be used locally without installation.  This is
useful for testing pre-release versions of the module.

Change-Id: I35d3a1009b677ef9d29379147312abe3d0a7f8b2