ci: Commit package-lock.json, use npm ci (#112)
It is a best practice for CI to commit package-lock.json and to use
npm ci instead of npm install, for consistent results in CI systems.

Package details were removed from package.json to prevent accidental
publication to npm.  This projects uses node modules for testing, but
is not itself a node module.