feat: Do not use libaom options, prefer libsvt-av1
libsvt-av1 can encode in real time with no special options, and does
not recognize the options we used in libaom to get as little as 0.2x
encoding in our "Player History" live stream.

From now on, let's assume that anyone who wants to use AV1 is using
libsvt-av1, which seems to be the only reasonable encoder.  If someone
is still using libaom, it will still work, but it will be using slower
default settings.