test: Fix WebM filter test (#155)
We no longer output opus in WebM, but rather in MP4. Therefore the WebM
filter test did not work any more when checking for opus. Instead, it
must check for vp9, which we still put into WebM.

This went unnoticed because there was no GitHub Actions workflow to run
the tests.