fix(cloud): Retry on all failures
Uses basic retry settings for GCS and S3.  If we do not pass these params, GCS would normally only retry if certain params were passed to convince it that the operation was idempotent.  We know our use case, so we know we can safely retry.  There will never be two versions of the same segment file, or two concurrent writes to the MPD, for example.