fix(cloud): Retry on all failures (#165)
Uses basic retry settings for GCS and S3.

If we do not pass these params, GCS would normally only retry if certain params were passed so that it knew the operation was definitely idempotent.  We know our use case, so we know we can safely retry.

There will never be two versions of the same segment file, or two concurrent writes to the MPD, for example.  So both uploads and deletes can be safely retried.