test: Fix mypy types (#169)
On some versions, mypy thinks the file handle from the web server is BinaryIO, in others, BufferedIOBase.  We don't care which it is, as both have the methods we care about.

I tried the IOBase class from "io", but I get failures on certain devices.  This is the best solution I've found so far.