fix(binaries): Bring PyPi package size under the limit (#192)
According to https://pypi.org/help/#file-size-limit, there is a limit of 100MB for any given wheel.  Our Linux x64 wheel is 101MB, and so the 1.0.0 release failed to upload.

This reduces the binary size by avoiding distro-specific versions of ffprobe.  The only thing we get from distro-specific versions of ffmpeg is hardware-encoding support, but that doesn't apply to ffprobe.  So it was wasteful to include distro-specific ffprobe builds.

This reduces the Linux x64 wheel size from 101MB to 69MB.