fix(docs): Improve READMEs and pypi docs for binaries (#214)
The binary package on pypi has almost no information.  This adds a more detailed README to make the pypi landing page more explanatory.

This also improves the main README to mention the binary package.  It is strongly recommended, and should be mentioned in the brief instructions on the main package README.

Classified as a "fix" only so that the release workflow will trigger, so that we can update the docs on pypi with a new release.