ci: Switch pull_request_target to pull_request
With this change, GitHub will start to enforce "require approval" settings for the repo, which `pull_request_target` allows PR authors to get around.

To stop using `pull_request_target`, we also have to stop using `vars`, which are not available on `pull_request` triggers.  Instead, we use the "environments" trick used by Shaka Packager's workflows before `vars` was introduced to GitHub Actions.  This is a safer system.